requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
httpx==0.26.0
starlette>=0.36.3 